% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_graph.R
\name{grid_to_graph}
\alias{grid_to_graph}
\title{Convert grid to proximity graph}
\usage{
grid_to_graph(grid)
}
\description{
Convert grid to proximity graph
}
\examples{
grid <- as.matrix(read.table(textConnection("
0 0 0 0 0 0
1 1 0 0 1 0
1 1 0 0 1 0
1 1 1 1 0 0
0 0 0 0 0 0
")))

g <- grid_to_graph(grid)
ssi(g$vertices, g$edges, group = 1)

}
