% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk_segregation.R
\name{random_walk_segregation}
\alias{random_walk_segregation}
\title{Random Walk Segregation Index}
\usage{
random_walk_segregation(vertices, edges, alpha = 0.85,
  edges_weight = "weight", vertices_id = 1, vertices_group = 2,
  verbose = 0)
}
\description{
Random Walk Segregation Index
}
