% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssi.R
\name{ssi}
\alias{ssi}
\title{Spectral Segregation Index}
\usage{
ssi(vertices, edges, group, edges_weight = "weight", vertices_id = 1,
  vertices_group = 2, verbose = 0)
}
\description{
Spectral Segregation Index
}
